[test] added seed scripts to app schema and test schema
